Add computed tests
