+++ packages/dx-grid-core/src/plugins/local-selection/computeds.js
@@ -1,6 +1,9 @@
-export const selectAllAvaliable = (selection, availableToSelect) => {
-  const availableToSelectSet = new Set(availableToSelect);
-  return selection.filter(selected => availableToSelectSet.has(selected)).length > 0;
+export const getAvailableToSelect = (rows, getRowId, isGroupRow) => {
+  let dataRows = rows;
+  if (isGroupRow) {
+    dataRows = dataRows.filter(row => !isGroupRow(row));
+  }
+  return dataRows.map(row => getRowId(row));
 };
 
 export const isSomeSelected = ({ selection, availableToSelect }) => {
