+++ packages/dx-grid-core/src/plugins/selection-state/computeds.js
@@ -1,11 +1,3 @@
-export const getAvailableToSelect = (rows, getRowId, isGroupRow) => {
-  let dataRows = rows;
-  if (isGroupRow) {
-    dataRows = dataRows.filter(row => !isGroupRow(row));
-  }
-  return dataRows.map(row => getRowId(row));
-};
-
 export const getAvailableSelection = (selection, availableToSelect) => {
   const availableToSelectSet = new Set(availableToSelect);
   return selection.filter(selected => availableToSelectSet.has(selected));
